Create directories if they don't exist

+++ test_driver.sh
@@ -1,16 +1,29 @@
 #! /bin/sh
+set -e
+if [ ! -d "/home/jovyan/opensarlab-notebook_testing/notebook_testing_logs" ]; then
+  mkdir /home/jovyan/opensarlab-notebook_testing/notebook_testing_logs
+  mkdir /home/jovyan/opensarlab-notebook_testing/notebook_testing_logs/old_logs
+fi
+if [ ! -d "/home/jovyan/opensarlab-notebook_testing/notebook_testing_dev" ]; then
+  mkdir /home/jovyan/opensarlab-notebook_testing/notebook_testing_dev
+fi
+if [ ! -d "/home/jovyan/opensarlab-notebook_testing/reports" ]; then
+  mkdir /home/jovyan/opensarlab-notebook_testing/reports
+fi
 oldlogs="/home/jovyan/opensarlab-notebook_testing/notebook_testing_logs/old_logs/*.log"
-for f in $oldlogs
-do
-   rm $f
-done
-# Re-install and activate rtc_analysis conda environment
+if [ -f "$oldlogs" ]; then
+   for f in $oldlogs
+   do
+      rm $f
+   done
+fi
+## Re-install and activate rtc_analysis conda environment
 conda init bash
 source ~/.bashrc
 conda env create -f '/home/jovyan/conda_environments/Environment_Configs/rtc_analysis_env.yml' --prefix "/home/jovyan/.local/envs/rtc_analysis" --force
 conda run -n rtc_analysis kernda --display-name rtc_analysis -o /home/jovyan/.local/envs/rtc_analysis/share/jupyter/kernels/python3/kernel.json
 conda activate /home/jovyan/.local/envs/rtc_analysis
-# Run tests on notebooks that use rtc_analysis environment
+## Run tests on notebooks that use rtc_analysis environment
 python /home/jovyan/opensarlab-notebook_testing/Test_Ecosystems_Ex1.py
 python /home/jovyan/opensarlab-notebook_testing/Test_Ecosystems_Ex2.py
 python /home/jovyan/opensarlab-notebook_testing/Test_Ecosystems_Ex3.py
@@ -30,21 +43,21 @@ python /home/jovyan/opensarlab-notebook_testing/Test_Master_Prepare_Data_Stack_H
 python /home/jovyan/opensarlab-notebook_testing/Test_Master_Subset_Data_Stack.py
 python /home/jovyan/opensarlab-notebook_testing/Test_Hazards_Prepare_Data_Stack_Hyp3_no_group.py
 python /home/jovyan/opensarlab-notebook_testing/Test_Hazards_Subset_Data_Stack.py
-# Deactive rtc_analysis environment
+## Deactivate rtc_analysis environment
 conda deactivate
-# Re-install and activate insar_analysis conda environment
+##Re-install and activate insar_analysis conda environment
 conda env create -f '/home/jovyan/conda_environments/Environment_Configs/insar_analysis_env.yml' --prefix "/home/jovyan/.local/envs/insar_analysis" --force
 conda run -n insar_analysis kernda --display-name insar_analysis -o /home/jovyan/.local/envs/insar_analysis/share/jupyter/kernels/python3/kernel.json
 source install_insar_analysis_pkgs.sh
 conda activate /home/jovyan/.local/envs/insar_analysis
-# Run tests on notebooks that use insar_analysis environment
+## Run tests on notebooks that use insar_analysis environment
 python /home/jovyan/opensarlab-notebook_testing/Test_Master_InSAR_volcano_source_modeling.py
 python /home/jovyan/opensarlab-notebook_testing/Test_GEOS657_Lab6.py
-# Deactive insar_analysis environment
+## Deactive insar_analysis environment
 conda deactivate
 echo 'Done running tests on the notebooks!'
 echo 'Check the logs for exceptions and failures!'
-# Get rid of spaces in log filenames
+## Get rid of spaces in log filenames
 oldnames="/home/jovyan/opensarlab-notebook_testing/notebook_testing_logs/*.log"
 for f in $oldnames
 do
@@ -90,6 +103,6 @@ for f in $infologs
 do
    mv $f "/home/jovyan/opensarlab-notebook_testing/notebook_testing_logs/old_logs/"
 done
-#mail -s "OSL Notebook Test Report" testern@alaska.edu <  test_report_$now.txt
+##mail -s "OSL Notebook Test Report" testern@alaska.edu <  test_report_$now.txt
 mv test_report_$now.txt "/home/jovyan/opensarlab-notebook_testing/reports/"
 echo 'Issues: '$issue_counter
